pyproject.toml
[project]
name = "civitai-api"
version = "2.0.0"
description = "Default template for PDM package"
authors = [
    {name = "Pyon~ Pyon~~", email = "32931796+tick97115115@users.noreply.github.com"},
]
dependencies = [
    "httpx[socks]>=0.28.1",
 "pydantic>=2.10.4",
 "anyio>=4.7.0",
 "mypy>=1.15.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "trio>=0.28.0",
    "python-dotenv>=1.0.1",
]

